Updated to use MDX syntax
