+++ packages/terra-base/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Updated docs to use mdx syntax
 
 5.25.0 - (January 7, 2020)
 ------------------
