+++ packages/terra-base/src/terra-dev-site/doc/base/Base.1.doc.mdx
@@ -1,3 +1,11 @@
+import { Badge } from '../../../../package.json?dev-site-package';
+
+import GlobalBaseStylesExample from './GlobalBaseStylesExample?dev-site-example'
+
+import BasePropsTable from '../../../../src/Base?dev-site-props-table';
+
+<Badge />
+
 # Terra Base
 
 The `terra-base` component handles two main concerns, [application internationalization](https://engineering.cerner.com/terra-ui/#/getting-started/terra-ui/internationalization/internationalization-intro) and global base styles.
@@ -77,3 +85,10 @@ const locale = (navigator.languages && navigator.languages[0])
 * [Localization Support](https://github.com/cerner/terra-ui/blob/master/src/terra-dev-site/contributing/ComponentStandards.e.contributing.md#internationalization-i18n-support)
 * [react_on_rails Compatible](https://github.com/shakacode/react_on_rails/blob/8cb06ed35cb5c2c453bcc193282b4c091574c1b7/docs/basics/i18n.md#how-to-add-i18n)
 * [CND Compatible](https://github.com/webpack/docs/wiki/configuration#outputpublicpath)
+
+
+## Example
+<GlobalBaseStylesExample />
+
+## Base Props
+<BasePropsTable />
