+++ packages/terra-base/src/terra-dev-site/doc/base/ChangeLog.2.doc.mdx
@@ -1,0 +1,3 @@
+import ChangeLog from '../../../../CHANGELOG.md';
+
+<ChangeLog />
